Allow hashing `State`.
This has a particularly subtle wrinkle: we consider "undefined value" to be equal to "default value", and we cannot enumerate all keys. Ergo, we need to ignore default values in the state when hashing, so that hashes of a state with a default value and a state with no value will compare equal.