Allow hashing and "ordering" of `Symbolic` values.
Note that the "ordering" provided here violates the semantics of symbolic execution, and exists purely so that types that need *an* ordering get one. If you need to symbolically order symbolic values, you will need to symbolically subtract them and analyze the result to determine if it's concretely-greater, concretely-less, concretely-equal, or unknown.