Consolidate the integer, signed-integer, float, and pointer-value parameters of `Literal` into a single new trait, removing the existing generic version in favor of assembler-specific versions.
Each assembler now has it's own `Literal` type that specifies what types it takes. Due to the need to still work with pointer values, we currently have utilites to extract them from a `Literal`, which means each assembler needs to have a preferred pointer size. This isn't ideal, it precludes assemblers that work with multiple architectures, and complicates trait bounds that I'm trying to simplify. Still, it works for now.