Make `dis` consider backreferences when searching for code that is part of a subroutine.
We don't check backreferences if the currently considered block has subroutine references into it. This is because some architectures make it very feasible to replace a call and return with an unconditional jump to the called function; in this case, we don't want to show every caller as part of the function being called.