Fix views starting from several lines into an instruction skipping a handful of instructions after it.
So, let me walk you through another weird consequence of how our UI code works. Our disassembler works on memory addresses, and they kind of have to. However, our UI works in lines, which are then grouped into addresses, optionally snapped to known instruction boundaries, and then grouped agan into regions. This introduces some complexity, most of which is packaged up into `Tumbler`.

However, there are still some residual problems. Notably, when rendering the disassembly view, we're suddently working in addresses rather than lines. This means that if our scroll tumbler is halfway through the current instruction, then we need to skip a few lines before we start rendering. Likewise, we need to scroll by the number of lines we did render.

Our existing code counted newlines in order to do both things. This is actually wrong. The number of lines to skip is actually the inverse of the number of lines we printed. For various programmer convenience reasons, we actually just calculate the lines to skip when we're getting our disassembly rendered, since that's when we have the memory address we need to count lines with.