Move the with_db/context macros to their own separate modules.
The any-database and any-context traits have been modified to support their use with generic architecture traits, which means there's less traits to modify if we add a new architecture.

This also means that I have to remove the SFC/65C816 names from the platform/arch traits. They'll be back, eventually.