Move the placeholder status of a symbol out of it's label and into the symbol itself.
This removes a handful of hacks to deal with the fact that a plain string would be different based on it's source.

Furthermore, it allows us to implement a more fine-grained control for symbol creation. We introduce a notion of "symbol strength" based on the source of a symbol. If a symbol comes from a subroutine reference, for example, then a new data reference to that particular symbol does NOT change the symbol's name if they share a label OR a pointer address. Ergo, autogenerated labels won't get repointed or renamed into weaker forms. User-specified labels are treated as the strongest type of symbol, and we also allow multiples of those as they may be coming from an assembler data source.